+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/extension/WebExtensionPromptFeature.kt
@@ -161,6 +161,7 @@ class WebExtensionPromptFeature(
                 return
             }
 
+            is WebExtensionInstallException.AdminInstallOnly,
             is WebExtensionInstallException.UnsupportedAddonType,
             is WebExtensionInstallException.Unknown,
             -> {