+++ mobile/android/fenix/app/src/test/java/org/mozilla/fenix/extension/WebExtensionPromptFeatureTest.kt
@@ -325,4 +325,22 @@ class WebExtensionPromptFeatureTest {
 
         verify { webExtensionPromptFeature.showDialog(expectedTitle, expectedMessage) }
     }
+
+    @Test
+    fun `WHEN calling handleInstallationFailedRequest with AdminInstallOnly error THEN showDialog with the correct message`() {
+        val expectedTitle = ""
+        val extensionName = "extensionName"
+        val exception = WebExtensionInstallException.AdminInstallOnly(
+            extensionName = extensionName,
+            throwable = Exception(),
+        )
+        val expectedMessage =
+            testContext.getString(R.string.mozac_feature_addons_failed_to_install, extensionName)
+
+        webExtensionPromptFeature.handleInstallationFailedRequest(
+            exception = exception,
+        )
+
+        verify { webExtensionPromptFeature.showDialog(expectedTitle, expectedMessage) }
+    }
 }